Port ZOL assembler framework.
This is a mess, it requires clang to compile the files that contain
assembler, or, inline assembler. So there is two paths for
headers to go through to try to accomodate it. But it does let us
use the Unix assembler work without much work.

Currently fletcher appears to run, but AES corrupts the stack,
the .S need the correct registers, which is supposed to be
rcx, rdx, r8, r9 on Win64 - and that appears to work, both
aes_setkey and aes_encrypt complete, and appear to stay inside
the memory passed along. However, something else may destroy the stack.